Catch case when no task in cluster
When it's the first time running the provision it runs `stopTask` to check if there exists any running tasks in the cluster, but since it's the first time, it returns an empty array `taskArns":[]`. Currently fails because the next function `getTasksDetails` will get an empty array and needs at least one task to function properly.